Split Hot.test.ts into two files
This halves run time.
